+++ src/main/java/com/atlauncher/gui/tabs/ServersTab.java
@@ -156,6 +156,11 @@ public class ServersTab extends JPanel implements Tab, RelocalizationListener {
     }
 
     @Override
+    public String getAnalyticsScreenViewName() {
+        return "Servers";
+    }
+
+    @Override
     public void onRelocalization() {
         clearButton.setText(GetText.tr("Clear"));
         searchButton.setText(GetText.tr("Search"));