+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Inject Lets Encrypt certificate on Java 8 < 141
 - Not all Forge libraries using our mirror
 - If the Java runtime doesn't exist for a Minecraft version, force refresh them
+- Issue with collapsing/expanding instances lagging when there's more than a couple
 
 ### Misc
 - Clean up Java utility class to clean up references to "Minecraft" Java concept