+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.10.1
+## 3.4.10.2
 
 ### New Features
 
 ### Fixes
-- Analytics events for new pack browser not being fully correct
-- Analytics events for tabs being translated and not being correct for tabs within tabs
 
 ### Misc